+++ example/ios/Runner/AppDelegate.swift
@@ -58,7 +58,7 @@ import flutter_callkit_incoming
     
     // Handle incoming pushes
     func pushRegistry(_ registry: PKPushRegistry, didReceiveIncomingPushWith payload: PKPushPayload, for type: PKPushType, completion: @escaping () -> Void) {
-        
+        print("didReceiveIncomingPushWith")
         guard type == .voIP else { return }
         
         let id = payload.dictionaryPayload["id"] as? String ?? ""